+++ test/e2e/fastify.e2e-spec.ts
@@ -2,13 +2,13 @@ import { resolve } from 'node:path'
 import process from 'node:process'
 
 import fastifyMultipart from '@fastify/multipart'
-import { afterAll, beforeAll, describe, expect, it } from '@jest/globals'
-
 import { ValidationPipe } from '@nestjs/common'
+
 import { FastifyAdapter } from '@nestjs/platform-fastify'
 import { Test } from '@nestjs/testing'
 import { fastify } from 'fastify'
 import { copy } from 'fs-extra'
+import { afterAll, beforeAll, describe, expect, it } from 'vitest'
 import type { NestFastifyApplication } from '@nestjs/platform-fastify'
 import type { TestingModule } from '@nestjs/testing'