+++ test/e2e/users.e2e-spec.ts
@@ -1,13 +1,13 @@
 import { resolve } from 'node:path'
 import process from 'node:process'
 
-import { afterAll, afterEach, beforeAll, beforeEach, describe, expect, it } from '@jest/globals'
 import { ValidationPipe } from '@nestjs/common'
-
 import { FastifyAdapter } from '@nestjs/platform-fastify'
+
 import { Test } from '@nestjs/testing'
 import { copy, readJson, writeJson } from 'fs-extra'
 import { authenticator } from 'otplib'
+import { afterAll, afterEach, beforeAll, beforeEach, describe, expect, it } from 'vitest'
 import type { NestFastifyApplication } from '@nestjs/platform-fastify'
 import type { TestingModule } from '@nestjs/testing'